requirements.txt
# === CORE FLASK FRAMEWORK ===
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Werkzeug==2.3.7

# === DATABASE ===
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
alembic==1.16.5

# === AUTHENTICATION & SECURITY ===
bcrypt==4.1.2
email-validator==2.2.0
WTForms==3.1.1
Flask-JWT-Extended==4.6.0

# === SCHEDULING & BACKGROUND TASKS ===
APScheduler==3.11.0
Flask-APScheduler==1.13.1

# === UTILITIES ===
python-dateutil==2.8.2
python-dotenv==1.0.0
Markdown==3.5.1

# === REPORTS & EXPORTS ===
reportlab==4.0.4
pandas==2.1.1
openpyxl==3.1.2

# === WEB SERVER (PRODUCTION) ===
gunicorn==23.0.0

# === ADDITIONAL DEPENDENCIES ===
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
greenlet==3.2.3
typing_extensions==4.14.0
six==1.17.0
pytz==2023.3
tzdata==2025.2

# === SYSTEM MONITORING ===
psutil==7.1.0

# === DEVELOPMENT & TESTING ===
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
flake8==6.1.0
black==23.11.0
isort==5.12.0

# === DOCUMENTATION ===
mkdocs==1.5.3
# mkdocs-material==9.4.8  # Opcional - requer instalação separada se desejar tema Material
# pymdown-extensions==10.3  # Opcional - extensões avançadas de Markdown

# === OPTIONAL DEPENDENCIES ===
Pillow==10.1.0
qrcode==7.4.2
celery==5.3.4
redis==5.0.1